requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
cryptography==43.0.0
bcrypt==4.1.3
python-dotenv==1.0.1
Werkzeug==3.0.3
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.4
pyotp==2.9.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
email-validator>=2.1.0
